refactor(cli): extract startup authentication
