+++ packages/agent-cli/agent-cli.cabal
@@ -59,6 +59,7 @@ library
                     , Agent.CLI.Session.Selection
                     , Agent.CLI.Session.Runtime.Types
                     , Agent.CLI.SessionAdmin
+                    , Agent.CLI.Startup.Auth
                     , Agent.CLI.StartupContext
                     , Agent.CLI.Startup.Format
                     , Agent.CLI.TUI.Composer.Buffer
